Switched all SIMTOI values over to doubles.  LibOI and OpenGL still use floats, downcasting happens as the model/shader/OpenCL level.

+++ src/CGLShader.h
@@ -59,7 +59,7 @@ public:
 
 	void LinkProgram(GLuint program);
 
-	void UseShader(float * params, unsigned int in_params);
+	void UseShader(double * params, unsigned int in_params);
 
 };
